C7FP-38 update gatling fix bug
