+++ gatling/gatling-50-artist.sh
