--- load_test/gatling/.gitignore
@@ -1,2 +1,0 @@
-results
-target
