--- load_test/gatling/resources/artist.csv
@@ -1,1 +1,0 @@
-Artist,Name,year,UUID
