--- load_test/gatling/resources/music.csv
@@ -1,7 +1,0 @@
-Artist,SongTitle,UUID
-Taylor Swift,The Last Great American Dynasty,6ecfafd0-8a35-4af6-a9e2-cbd79b3abeea
-Backxwash,Bad Juju,c2573193-f333-49e2-abec-182915747756
-Sini Sabotage,Mun Planeeta,22e47f97-11ca-4c3c-8e77-f3068fddaf6e
-William Prince,Reliever,26e6462c-eacf-40bb-b4d0-d683966e2624
-Gang of Four,Isle of Dogs,bb9e9105-6224-4a60-aabd-8d29cc4b1349
-Carla Bley Trio,Utviklingssang,0894ddc7-0c84-4f13-a037-ddcaa1134ec8
