--- load_test/gatling/resources/users.csv
@@ -1,5 +1,0 @@
-fname,lname,email,UUID
-George,Clinton,starchild@pfunk.org,b66f61e7-4fad-402f-aaad-101e3e2522d3
-Tina,Weymouth,onlytina@tomtomclub.com,75ebc062-a808-4a0f-a79b-8ac568564bfe
-Louis,Armstrong,satchmo@nola.gov,2e4c0d7e-b0e6-4739-b044-ee1d45872a7c
-Leon,Beiderbecke,bix@aa.org,423a10a6-ab66-48c5-a1c7-dffb3169d744
